Protect against NPE when glom.document.locale is not in config.
This patch protects against an NPE when glom.document.locale is not in
the config file. This NPE will also happen if glom.document.locale is
commented out.

The patch also updates the error message to display the class name when
the getMessage() returns null. This was happening when the NPE was
thrown and I had "Configuration Error: null". If an NPE is encountered
with this patch, "Configuration Error: NullPointerException " will be
displayed.

This commit closes this bug:

https://bugzilla.gnome.org/show_bug.cgi?id=666669

* src/main/java/org/glom/web/server/OnlineGlomServiceImpl.java: