DetailsPlace: Use a date-formatting API available in GWT client code.
        * src/main/java/org/glom/web/client/place/DetailsPlace.java:
        getStringForData(): Use DateTimeFormat.getFormat(), because that works
on the client side (but not the server side) instead of String.format()
with Date.getYear()/getMonth()/getDate(), because they work only on the
server side (but not on the client side). Rather annoying.
        * src/test/java/org/glom/web/client/place/DetailsPlaceTest.java: Rename to
GwtTestDetailsPlace.
        * src/test/java/org/glom/web/client/place/ListPlaceTest.java: Rename to
GwtTestListPlaceTest, so we can run these as client-side (generated
Javascript) code so we can catch these problems sooner.
It is unfortunate that the code cannot run on the server-side as Java,
but it does not need to.